+++ CHANGELOG.md
@@ -8,6 +8,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ### Add
 - Tabler icons catalog.
 - MingCute icons catalog.
+- Remix icons catalog.
 
 ### Fixed
 - Updated icons: `lucide`, `simple-icons`, `material-symbols`.